requirements.txt
astroid==2.12.10
attrs==22.1.0
bcrypt==3.2.0
certifi==2022.6.15
cffi==1.14.5
charset-normalizer==2.1.1
click==7.1.2
dill==0.3.5.1
distlib==0.3.6
dnspython==2.2.1
exceptiongroup==1.0.0rc9
filelock==3.8.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.7.0
packaging==21.3
Pillow==9.3.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2==2.9.2
psycopg2-binary==2.9.5
pycparser==2.20
pygame==2.1.2
pylint==2.15.3
pymongo==4.3.3
pyparsing==3.0.9
pytest==7.2.0
python-dotenv==0.15.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.3.23
SQLAlchemy-Utils==0.36.8
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
urllib3==1.26.12
virtualenv==20.16.3
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
zipp==3.7.0
